backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
httpx
cryptography
python-telegram-bot[job-queue]>=21.0
